% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kbinary}
\alias{kbinary}
\title{K-binary clustering}
\usage{
kbinary(x, k, nstart = 1, trace = FALSE)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{k}{The number of clusters to be considered. A random set of (distinct) rows in x is chosen as the initial centres.}

\item{nstart}{Number of random sets that has been chosen}

\item{trace}{logical: if true, tracing information on the progress of the algorithm is produced.}
}
\value{
\itemize{
  \item A - The letters of the alphabet.
  \item B - A vector of numbers.
}
}
\description{
Perform k-binary clustering
}
