% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kbinary_gibbs}
\alias{kbinary_gibbs}
\title{K-dissimilarities algorithm with uncertainty quantification}
\usage{
kbinary_gibbs(
  x,
  k,
  lambda = 1,
  R = 1000,
  burn_in = 1000,
  nstart = 10,
  trace = FALSE
)
}
\arguments{
\item{x}{numeric matrix of of the data}

\item{k}{The number of clusters to be considered.}

\item{R}{Number of MCMC samples after burn-in}

\item{burn_in}{Number of MCMC samples to be discarded as burn-in period}

\item{nstart}{Number of random initializations for the k-means algorithm}

\item{trace}{logical: if true, tracing information on the progress of the algorithm is produced.}
}
\value{
\describe{
  \item{\code{G}}{The letters of the alphabet}
  \item{\code{lambda}}{A vector of numbers}
  \item{\code{loss}}{A vector of numbers}
  \item{\code{G_map}}{A vector of numbers}
  \item{\code{loss_map}}{A vector of numbers}
}
}
\description{
Perform the Gibbs-sampling for the k-dissimilarities algorithm using the Minkowski distance; see \code{\link{dist}}.
}
