% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kbinary_select}
\alias{kbinary_select}
\title{Selection of the number of cluster for the k-binary algorithm}
\usage{
kbinary_select(x, k_max, nstart = 1)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{k_max}{The maximum number of clusters to be considered. A random set of (distinct) rows in x is chosen as the initial centres.}

\item{nstart}{Number of random sets that has been chosen}
}
\value{
It plots the loss function for different clustering solutions
}
\description{
It displays the value of the loss function for various choices of k.
}
