% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kdiss_select}
\alias{kdiss_select}
\title{Selection of the number of cluster for the k-dissimilarities algorithm}
\usage{
kdiss_select(D, k_max, nstart = 1, method = "elbow")
}
\arguments{
\item{D}{A \code{n x n} numeric matrix with the dissimilarities, typically the output of \code{\link{dist}} or \code{\link{daisy}}.}

\item{k_max}{Maximum number of clusters to be considered.}

\item{nstart}{Number of random initializations.}

\item{method}{The graph that will be displayed. Supported options are \code{method="elbow"}, which displays the loss function, or \code{method="silhouette"}. See \code{\link{silhouette}} for details about the latter.}
}
\value{
It return a \code{\link{ggplot2}} graph of the loss function / average silhouette width, for \code{k=1,...,k_max}.
}
\description{
It displays the value of the loss function / average silhouette width, for different values of \code{k}
}
