% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{comp_medoids}
\alias{comp_medoids}
\title{Computation of the medoids}
\usage{
comp_medoids(D, cluster)
}
\arguments{
\item{D}{A \code{n x n} numeric matrix with the dissimilarities, typically the output of \code{\link{dist}} or \code{\link{daisy}}.}

\item{cluster}{A clustering solution, typically the output of \code{\link{kdiss}}.}
}
\value{
\describe{
  \item{\code{medoids}}{Labels of the medoids}
}
}
\description{
Compute the medoids of a given clustering solution.
}
