% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kdiss}
\alias{kdiss}
\title{K-dissimilarities clustering}
\usage{
kdiss(D, k, nstart = 1, trace = FALSE)
}
\arguments{
\item{D}{A n x n numeric matrix with the dissimilarities}

\item{k}{The number of clusters to be considered}

\item{nstart}{Number of random initializations}

\item{trace}{logical: if true, tracing information on the progress of the algorithm is produced}
}
\value{
\describe{
  \item{\code{cluster}}{The letters of the alphabet}
  \item{\code{loss}}{A vector of numbers}
}
}
\description{
Perform the k-dissimilarities Algorithm described in Rigon, Herring and Dunson (2020).
}
