% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kdiss_select}
\alias{kdiss_select}
\title{Selection of the number of cluster for the k-dissimilarities algorithm}
\usage{
kdiss_select(D, k_max, nstart = 1, method = "elbow")
}
\arguments{
\item{D}{A n x n numeric matrix with the dissimilarities}

\item{k_max}{Maximum number of clusters to be considered}

\item{nstart}{Number of random initializations}

\item{method}{The graph that has to reported, i.e. \code{elbow} or \code{silhouette}}
}
\value{
It return a \code{ggplot2} graph of the loss function / average silhouette width, for different values of \code{k}, from 1 up to \code{k_max}.
}
\description{
It displays the value of the loss function / average silhouette width, for different values of \code{k}
}
